Define legendLabels in SingleRoomTimeline
